Time: 0 ms (100%), Space: 7.8 MB (87.8%) - LeetHub

+++ 0070-climbing-stairs/0070-climbing-stairs.cpp
@@ -4,7 +4,7 @@ public:
     
 
     int climbStairs(int n) {
-        long int answer = 0 , temp1 = 1 , temp2 = 2 ;
+         int answer = 0 , temp1 = 1 , temp2 = 2 ;
 
         if(n==1) return 1;
         if(n==2)return 2;
